requirements.txt
# Core dependencies
yt-dlp>=2023.12.30
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0

# Instagram API
instagrapi>=2.0.0

# AI Caption Generation
openai>=1.0.0
anthropic>=0.7.0

# Scheduling
APScheduler>=3.10.4

# Database
sqlalchemy>=2.0.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Video Processing
ffmpeg-python>=0.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0

# Logging
colorlog>=6.7.0

# Content Safety
better-profanity>=0.7.0
langdetect>=1.0.9
